requirements.txt
pyenchant
pytest
termcolor
